requirements.txt
pandas
numpy
scikit-learn
xgboost
statsmodels
joblib
fastapi[standard]  # Keep this if you need FastAPI's standard dependencies
pyyaml
flask
mlflow
dagshub